Add function and unit test to return name and milileage data of each participant
Function visits each region's webpage (each region has a different webpage),
scrapes particpant data, and returns the data in a dictionary where region number
is the key and a dict of participant+ miles logged for that region is the value.
The function also returns a set of the participants in the race. It requires some NER.