Update get_participant_data so that it returns participants' ID info
Previous function returned a set of all participants in the event and a nested dict
representing which people participated in each region, and how many miles each
contributed for each region. The updated funtion scrapes and returns each participant's
sex and age, along with name, in a tuple, too. This info will be used for some
entity resolution that needs to be done.

The unit test for this function was updated to reflect changes. Also, the test
is now a subtest now with each of the three data structures returned
tested separately.