Add function that makes HTTP call and gets back participant identification details
"get_identifiers" removes the need to scrape for participant name, gender, age, etc.
It accepts an HREF as an argument. It parses the query portion of the HREF, then
creates a new URL specific to the participant from the query details. The new URL
is used to make an HTTP request for participant identification details.

Additionally, the "get_participant_data" function has been refactored to use
"get_identifiers" and a unit test has been added to test it. This commit also
includes general refactoring of the test for "get_participant_data".