Refactor `get_participant_data`: function now collects all data via HTTP calls
Due to the new data source:
1) The exported Excel file now uses full names for each participant.
2) Unit tests have been updated to test for full names.

BS4 is still used in this funtion, but only to scrape for a name and an
HREF related to each participant. The HREF is used to make all HTTP calls
needed to gather data. The name is used alongside a `seen` dict to
ensure that unecessary HTTP calls within the function loops are avoided.