Add flask, add minimal frontend, and update filenames
note: script.py and script_test.p -> main.py and main_test.py

The frontend accepts user input, a team name. The program
uses the team name to generate and export a spreadsheet for this team.
Each function has been updated to accept this input and pass it down
in the data collection process.

Currently, the program works when the input is: "In Jesper's Footsteps".
It needs to be generalized to work with other team names, too.